requirements.txt
PyQt5==5.15.10
requests==2.31.0
Pillow==10.1.0
pyinstaller>=5.13.0


